+++ bot.js
@@ -1,6 +1,13 @@
+//===========================
+//Node.JS Version Check
+//===========================
+
 if (process.version.slice(1).split('.')[0] < 8) throw new Error('Node 8.0.0 or higher is required. Update Node on your system.');
 
+//===========================
 //Dependencies
+//===========================
+
 const Discord = require('discord.js');
 const chalk = require('chalk');
 const fs = require('fs');
@@ -13,14 +20,18 @@ const log = message => {
   console.log(`[${moment().format('YYYY-MM-DD HH:mm:ss')}] ${message}`);
 };
 
+//===========================
+//Command Handler
+//===========================
+
 client.commands = new Discord.Collection();
 client.aliases = new Discord.Collection();
 fs.readdir('./commands/', (err, files) => {
   if (err) console.error(err);
   log(`Loading a total of ${files.length} commands.`);
   files.forEach(f => {
-    let props = require(`./commands/${f}`);
-    log(`Loading Command: ${props.help.name}. 👌`);
+    const props = require(`./commands/${f}`);
+    log(`Loading Command - ${props.help.name}: ✅`);
     client.commands.set(props.help.name, props);
     props.conf.aliases.forEach(alias => {
       client.aliases.set(alias, props.help.name);
@@ -28,11 +39,15 @@ fs.readdir('./commands/', (err, files) => {
   });
 });
 
+//===========================
+//Reload Function
+//===========================
+
 client.reload = command => {
   return new Promise((resolve, reject) => {
     try {
       delete require.cache[require.resolve(`./commands/${command}`)];
-      let cmd = require(`./commands/${command}`);
+      const cmd = require(`./commands/${command}`);
       client.commands.delete(command);
       client.aliases.forEach((cmd, alias) => {
         if (cmd === command) client.aliases.delete(alias);
@@ -48,32 +63,39 @@ client.reload = command => {
   });
 };
 
+//===========================
+//User Permissions
+//===========================
+
 client.elevation = message => {
   let permlvl = 0;
-  let modRole = message.guild.roles.find('name', config.modrolename);
+  const modRole = message.guild.roles.find('name', config.modrolename);
   if (modRole && message.member.roles.has(modRole.id)) permlvl = 2;
-  let adminRole = message.guild.roles.find('name', config.adminrolename);
+  const adminRole = message.guild.roles.find('name', config.adminrolename);
   if (adminRole && message.member.roles.has(adminRole.id)) permlvl = 3;
   if (message.author.id === config.ownerid) permlvl = 4;
   return permlvl;
 };
 
-
-
+//===========================
 //Debugging and Error logging
-var regToken = /[\w\d]{24}\.[\w\d]{6}\.[\w\d-_]{27}/g; //Redacts token in debug logging.
+//===========================
 
+var regToken = /[\w\d]{24}\.[\w\d]{6}\.[\w\d-_]{27}/g; //Redacts token in debug logging.
 client.on('debug', e => {
-  console.log(e.replace(regToken, 'that was redacted'));
+  console.log(e.replace(regToken, (chalk.red('in config.json.'))));
+});
+
+client.emit('warn',e => {
+  console.log(chalk.bgYellow(e));
 });
 
 client.on('error', e => {
-  console.log(chalk.bgRed(e.replace(regToken, 'that was redacted')));
+  console.log(chalk.bgRed(e));
 });
 
-client.on('warn', e => {
-  console.log(chalk.bgYellow(e.replace(regToken, 'that was redacted')));
-});
-
+//===========================
 //Discord Login
+//===========================
+
 client.login(config.token);