+++ events/guildBanAdd.js
@@ -1,3 +1,6 @@
+const config = require('../config.json');
+
 module.exports = (guild, user) => {
-  guild.defaultChannel.send(`${user.tag} has been banned!`);
+  guild.channels.get(config.defaultChannel).send(`${user.tag} has been banned!`);
 };
+*/