+++ events/guildMemberAdd.js
@@ -1,4 +1,6 @@
+const config = require('../config.json');
+
 module.exports = member => {
-  let guild = member.guild;
-  guild.defaultChannel.send(`Please welcome ${member.user.username} to the server!`);
+  const guild = member.guild;
+  guild.channels.get(config.defaultChannel).send(`Please welcome ${member.user.username} to the server!`);
 };