+++ events/ready.js
@@ -1,5 +1,6 @@
 const chalk = require('chalk');
 
-module.exports = client => {
+module.exports = client => { //eslint-disable-line no-unused-vars
   console.log(chalk.bgGreen.black('I\'m online and ready to serve!'));
+  client.user.setGame('sybilserver.com');
 };